Bani 2.2 — trustworthy money: throttled bank sync, atomic restore, exact loan booking, Raport-first launch
Fixes all 12 findings of the 2026-08-29 full diagnostic (2 HIGH, 5 MED, 5 LOW) plus D-approved launch-tab flip.
- Bank sync: 6h foreground throttle (GoCardless quota-safe), visible last-sync/error, exact amount parsing
- Backup restore: atomic — a failed restore can no longer wipe data; blobs swapped only post-commit
- Loans: payments bound to exact schedule rows (out-of-order/edit-proof), double-tap-safe markDone, non-amortizing terms rejected, truncated cost flagged
- Reconciliation: unabsorbed residual recorded + shown; dedup merge repoints references
- FX: BNR rate as exact Decimal in every displayed total
- Launch tab: Raport-first (D sign-off)

+++ Bani/Views/Finances/Analytics/FinancesAnalytics.swift
@@ -35,11 +35,14 @@ enum FinancesAnalytics {
 
     /// The RON-equivalent of one item. `.ron` passes through; `.eur` converts
     /// with `rate` when present, otherwise returns the raw EUR amount (the caller
-    /// only combines when a rate exists — see `combinedTotal`).
-    static func ronValue(_ item: SpendItem, rate: Double?) -> Decimal {
+    /// only combines when a rate exists — see `combinedTotal`). L5: `rate` is a
+    /// `Decimal` (see `RateService.rateDecimal`) so this never re-derives a
+    /// `Decimal` from the display `Double` — the source of the FX display noise
+    /// this fix removes.
+    static func ronValue(_ item: SpendItem, rate: Decimal?) -> Decimal {
         switch item.currency {
         case .ron: return item.amount
-        case .eur: return rate.map { item.amount * Decimal($0) } ?? item.amount
+        case .eur: return rate.map { item.amount * $0 } ?? item.amount
         }
     }
 
@@ -55,11 +58,11 @@ enum FinancesAnalytics {
     /// RON-combined total over the items. With `rate == nil`, EUR items are
     /// EXCLUDED (there is no honest way to combine them) — the caller shows the
     /// side-by-side per-currency fallback instead.
-    static func combinedTotal(_ items: [SpendItem], rate: Double?) -> Decimal {
+    static func combinedTotal(_ items: [SpendItem], rate: Decimal?) -> Decimal {
         items.reduce(Decimal(0)) { sum, item in
             switch item.currency {
             case .ron: return sum + item.amount
-            case .eur: return rate.map { sum + item.amount * Decimal($0) } ?? sum
+            case .eur: return rate.map { sum + item.amount * $0 } ?? sum
             }
         }
     }
@@ -86,7 +89,7 @@ enum FinancesAnalytics {
     /// Per-category totals (RON-combined when `rate` given, else the passed items
     /// should already be single-currency), sorted by total descending. Custom
     /// categories aggregate as their own `CategoryRef.custom` buckets (C3).
-    static func byCategory(_ items: [SpendItem], rate: Double?) -> [CategoryTotal] {
+    static func byCategory(_ items: [SpendItem], rate: Decimal?) -> [CategoryTotal] {
         var totals: [CategoryRef?: (total: Decimal, count: Int)] = [:]
         for item in items {
             let value = ronValue(item, rate: rate)
@@ -121,7 +124,7 @@ enum FinancesAnalytics {
 
     /// Sums items into `unit`-aligned buckets across `interval`, filling empty
     /// buckets with zero so the bar chart never shows a gap (D2 empty states).
-    static func buckets(_ items: [SpendItem], interval: DateInterval, unit: Calendar.Component, calendar: Calendar, rate: Double?) -> [TimeBucket] {
+    static func buckets(_ items: [SpendItem], interval: DateInterval, unit: Calendar.Component, calendar: Calendar, rate: Decimal?) -> [TimeBucket] {
         var sums: [Date: Decimal] = [:]
         for item in items {
             let start = calendar.dateInterval(of: unit, for: item.date)?.start ?? item.date
@@ -169,7 +172,7 @@ enum FinancesAnalytics {
 
     /// A running-sum series for one window. `displayShift` maps a previous-period
     /// point onto the current window (0 for the current period).
-    static func cumulative(_ items: [SpendItem], rate: Double?, isPrevious: Bool, displayShift: TimeInterval) -> [CumulativePoint] {
+    static func cumulative(_ items: [SpendItem], rate: Decimal?, isPrevious: Bool, displayShift: TimeInterval) -> [CumulativePoint] {
         let sorted = items.sorted { $0.date < $1.date }
         var running = Decimal(0)
         return sorted.map { item in
